+++ src/components/about/about.tsx
@@ -7,7 +7,7 @@ import { makeStyles } from "@material-ui/core/styles"
 import Grid from "@material-ui/core/Grid"
 import Typography from "@material-ui/core/Typography"
 import { List } from "@material-ui/core";
-import AboutBulletPoint from "./AboutBulletPoint";
+import { BulletPoint } from "./components";
 
 
 const useStyles = makeStyles(theme => ({
@@ -20,6 +20,7 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
+
 function About() {
   const classes = useStyles();
 
@@ -30,7 +31,7 @@ function About() {
           <Typography className={classes.text} variant="h3">🦈 Sublet Shark</Typography>
         </Grid>
         <Grid item>
-          <Typography className={classes.text} variant="h5">Sublets for students</Typography> 
+          <Typography className={classes.text} variant="h5">Sublets for students</Typography>
         </Grid>
         <Grid item>
           <Grid container direction="row" alignItems="flex-start">
@@ -39,11 +40,11 @@ function About() {
                 Sublet Shark <strong>is</strong>:
               </Typography>
                 <List>
-                  <AboutBulletPoint 
+                  <BulletPoint 
                     checkMark 
                     text={"For post-secondary students looking for short term accommodation near their institution"} 
                   />
-                  <AboutBulletPoint
+                  <BulletPoint
                     checkMark 
                     text={"For post-secondary students looking to sublet their room/apartment to another local student"} 
                   />
@@ -54,9 +55,9 @@ function About() {
                 Sublet Shark <strong>is not</strong>:
               </Typography>
                 <List>
-                  <AboutBulletPoint text={"For leases"} />
-                  <AboutBulletPoint text={"For landlords or leasing agents, or anyone who isn't a student"} />
-                  <AboutBulletPoint text={"For long-term rental arrangements"} />
+                  <BulletPoint text={"For leases"} />
+                  <BulletPoint text={"For landlords or leasing agents, or anyone who isn't a student"} />
+                  <BulletPoint text={"For long-term rental arrangements"} />
                 </List>
             </Grid>
           </Grid>
