+++ src/components/about/components/bullet-point/index.tsx
@@ -1,0 +1,1 @@
+export { default as BulletPoint } from './bullet-point';
