+++ src/components/about/components/index.tsx
@@ -1,0 +1,1 @@
+export { BulletPoint } from './bullet-point'
