+++ src/components/about/index.tsx
@@ -1,0 +1,1 @@
+export { default as About } from './about';
