+++ src/components/index.tsx
@@ -1,0 +1,4 @@
+export { About } from './about';
+export { LoginForm } from './login-form';
+export { RegisterForm } from './register-form';
+export { RecoverPasswordForm } from './recover-password-form';
