+++ src/components/lock-text/index.tsx
@@ -1,0 +1,1 @@
+export { default as LockText } from './lock-text'
