+++ src/components/login-form/index.tsx
@@ -1,0 +1,1 @@
+export { default as LoginForm } from './login-form'
