+++ src/components/recover-password-form/index.tsx
@@ -1,0 +1,1 @@
+export { default as RecoverPasswordForm } from './recover-password-form';
