+++ src/components/register-form/index.tsx
@@ -1,0 +1,1 @@
+export { default as RegisterForm } from './register-form';
