+++ src/views/index.tsx
@@ -1,0 +1,3 @@
+export { LoginPage } from './login';
+export { RegisterPage } from './register';
+export { RecoverPasswordPage } from './recover-password';
