+++ src/views/login/index.tsx
@@ -1,0 +1,1 @@
+export { default as LoginPage } from './login';
