+++ src/views/recover-password/index.tsx
@@ -1,0 +1,1 @@
+export { default as RecoverPasswordPage } from './recover-password';
