+++ src/views/register/index.tsx
@@ -1,0 +1,1 @@
+export { default as RegisterPage } from './register';
